Moved locust tests and added a single action to debug
